test: add benchmarks for MergedFileDescriptors
Add a basic benchmark for MergedGlobalFileDescriptors, and add a new
package so that we can import jhump/protoreflect in order to dynamically
create new file descriptors. Right now, this is a crude implementation
that produces very simple file descriptors; but if we need to improve
the test cases, we now have a framework in place.