docs: updated installation instructions in README.md
as `README.md` states, this repository is a fork of
[gogo/protobuf](https://github.com/gogo/protobuf). the upstream
repository was [deprecated][deprecation-announce] in
[october, 2022][deprecation-commit].

there aren't any installation instructions for this repository written
down. rather, a reader must navigate to the [original readme] and
transpose the `go get` commands to instead point to this repository.

this copies the installation instructions over from the original readme,
updating the github url's to point here.

[original readme]: https://raw.githubusercontent.com/gogo/protobuf/master/Readme.md
[deprecation-announce]: https://twitter.com/awalterschulze/status/1584553056100057088
[deprecation-commit]: https://github.com/gogo/protobuf/commit/f67b8970b736e53dbd7d0a27146c8f1ac52f74e5